Comments addressed.
